JE2-89. Deleted not needed extra validation for unauthorized user in login form, moved login form logic to a separate hook
