+++ frontend/src/features/users/components/LoginForm.tsx
@@ -1,42 +1,10 @@
 import Grid from '@mui/material/Grid2'
 import { Box, Button, TextField, Typography } from '@mui/material'
-import React, { ChangeEvent, useState } from 'react'
-import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts'
-import { toast } from 'react-toastify'
-import { clearLoginError, selectLoginError, selectLoadingLoginUser } from '../../../store/slices/userSlice.ts'
-import { loginUser } from '../../../store/thunks/userThunk.ts'
 import { getFieldError } from '../../../utils/getFieldError.ts'
+import { useLoginForm } from '../hooks/useLoginForm.ts'
 
 const LoginForm = () => {
-  const dispatch = useAppDispatch()
-
-  const sending = useAppSelector(selectLoadingLoginUser)
-  const backendError = useAppSelector(selectLoginError)
-  const [form, setForm] = useState({ email: '', password: '' })
-
-  const onSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    try {
-      await dispatch(loginUser(form)).unwrap()
-      setForm({ email: '', password: '' })
-      dispatch(clearLoginError())
-      toast.success('Вы успешно вошли!')
-    } catch (error) {
-      if (error && Array.isArray(error)) {
-        const errorMessages = error.join('; ')
-        toast.error(errorMessages)
-      } else {
-        toast.error('Ошибка входа')
-      }
-    }
-  }
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    dispatch(clearLoginError(e.target.name))
-    setForm(data => ({ ...data, [e.target.name]: e.target.value.trim() }))
-  }
-
-  const isFormValid = form.email.trim() !== '' && form.password.trim() !== ''
+  const { form, handleChange, onSubmit, isFormValid, sending, backendError } = useLoginForm()
 
   return (
     <Box noValidate component="form" onSubmit={onSubmit} style={{ maxWidth: '20%', margin: '0 auto' }}>
